fix: move plugins api to plugins folder
