+++ catalog/internal/pluginmanager/plugins_test.go
@@ -5,8 +5,8 @@ import (
 	"net"
 	"testing"
 
-	"github.com/naira-project/naira/catalog/pluginapi"
-	pluginv1 "github.com/naira-project/naira/catalog/pluginapi/proto/plugin/v1"
+	pluginapi "github.com/naira-project/naira/plugins/pkg/api"
+	pluginv1 "github.com/naira-project/naira/plugins/pkg/api/proto/plugin/v1"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"google.golang.org/grpc"
