+++ plugins/cmd/litellm/plugin.go
@@ -9,7 +9,7 @@ import (
 	"net/http"
 	"strings"
 
-	"github.com/naira-project/naira/catalog/pluginapi"
+	pluginapi "github.com/naira-project/naira/plugins/pkg/api"
 	"k8s.io/client-go/dynamic"
 	"k8s.io/client-go/rest"
 )
