+++ plugins/pkg/api/grpc.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"fmt"
 
-	pluginv1 "github.com/naira-project/naira/catalog/pluginapi/proto/plugin/v1"
+	pluginv1 "github.com/naira-project/naira/plugins/pkg/api/proto/plugin/v1"
 )
 
 type GRPCServer struct {
