+++ plugins/pkg/api/proto/plugin/v1/plugin.proto
